project/project_1.0/requirements.txt
typer[all]
python-dotenv
neo4j
rich
